implementing post form component
